app/requirements.txt
fastapi
uvicorn
pydantic
requests
pymongo 
python-dotenv
redis
passlib[bcrypt]
pydantic[email]
flask-cors
motor
datetime
python-dateutil
openai